backend/requirements.txt
fastapi
uvicorn
linkup-sdk
langchain
langchain-community
langchain-ollama
langchain-anthropic
langgraph
sqlalchemy
alembic
playwright
beautifulsoup4
requests
pydantic
pydantic-settings
python-dotenv
lxml
html2text
pyyaml
tenacity
structlog
